requirements-test.txt
django-fake-model>=0.1.4